requirements.txt
accelerate==0.30.1
av==12.0.0
certifi==2024.2.2
charset-normalizer==3.3.2
diffusers==0.21.2
einops==0.8.0
filelock==3.14.0
fsspec==2024.5.0
huggingface-hub==0.23.0
idna==3.7
importlib_metadata==7.1.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
packaging==24.0
pillow==10.3.0
pkg_resources==0.0.0
psutil==5.9.8
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.1
safetensors==0.4.3
sympy==1.12
tokenizers==0.19.1
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
tqdm==4.66.4
transformers==4.41.0
triton==2.3.0
typing_extensions==4.11.0
urllib3==2.2.1
zipp==3.18.2
